--- src/reflect/components/observation_model/__init__.py
@@ -1,3 +1,0 @@
-from reflect.components.observation_model.observation_model import ObservationalModel
-from reflect.components.observation_model.encoder import ConvEncoder
-from reflect.components.observation_model.decoder import ConvDecoder
