+++ src/reflect/components/rssm_world_model/tests/test_rssm_world_model.py
@@ -1,7 +1,7 @@
 import torch
 from reflect.data.loader import EnvDataLoader
 from reflect.components.rssm_world_model.world_model import WorldModel
-from reflect.components.actor import Actor
+from reflect.components.models.actor import Actor
 from reflect.components.trainers.reward.reward_trainer import RewardGradTrainer
