+++ src/reflect/components/trainers/value/tests/test_value_rssm.py
@@ -1,6 +1,6 @@
 from reflect.data.loader import EnvDataLoader
 from reflect.components.rssm_world_model.world_model import WorldModel
-from reflect.components.actor import Actor
+from reflect.components.models.actor import Actor
 from reflect.components.trainers.value.value_trainer import ValueGradTrainer
