+++ src/reflect/components/transformer_world_model/transformer.py
