+++ src/reflect/components/transformer_world_model/world_model.py
@@ -1,9 +1,9 @@
 from typing import Optional
 from dataclasses import dataclass
-from reflect.components.observation_model.encoder import ConvEncoder
-from reflect.components.observation_model.decoder import ConvDecoder
-from reflect.components.transformer_world_model.backend.pytfex import PytfexTransformer
-from reflect.components.actor import Actor
+from reflect.components.models.encoder import ConvEncoder
+from reflect.components.models.decoder import ConvDecoder
+from reflect.components.transformer_world_model.transformer import PytfexTransformer
+from reflect.components.models.actor import Actor
 from itertools import chain
 from reflect.components.base import Base
 from reflect.utils import FreezeParameters
