+++ src/reflect/examples/differential_rl/value_trainer.py
@@ -4,7 +4,7 @@ from tqdm import tqdm
 
 from reflect.data.differentiable_pendulum import DiffPendulumEnv
 from reflect.components.trainers.value_trainer.value_trainer import ValueGradTrainer, update_target_network
-from reflect.components.actor import Actor
+from reflect.components.models.actor import Actor
 from reflect.components.trainers.value_trainer.value_critic import ValueCritic
 
 actor = Actor(
